Convert Twig functions to methods on the Twig extension
This allows twig to compile the calls to faster code, as it does not
need to load functions from all extensions to access the callable.
The implementation of the t() function has also been changed to collect
parameters at compile-time instead of using func_get_args in the
function (which does not always work because Twig validates the number
of arguments passed to the function when using named arguments).